backend/pyproject.toml
[project]
name = "mini-uploader"
version = "0.1.0"
description = "QR-gated image uploader that scans and forwards to Immich"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "pydantic-settings>=2.3",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "itsdangerous>=2.2",
    "bcrypt>=4.1",
    "cryptography>=42",
    "clamd>=1.0.2",
    "python-magic>=0.4.27",
]

[project.optional-dependencies]
test = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]
